feat(toast): refacto handle theme
