--- client/src/utils/initDarkMode.js
@@ -1,10 +1,0 @@
-
-export const initializeTheme = (isDarkMode) => {
-  const savedTheme = localStorage.getItem('theme');
-  if (savedTheme) {
-    isDarkMode.value = savedTheme === 'dark';
-  } else {
-    isDarkMode.value = window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches;
-  }
-  document.documentElement.classList.toggle('dark', isDarkMode.value);
-};
